Many additions:  - Materials  - Shader loading from files  - Color consts  - more...

+++ Include/C3CRC.h
@@ -1,7 +1,7 @@
 // **************************************************************
 // Celerity v3 Game / Visualization Engine Source File
 //
-// Copyright © 2001-2020, Keelan Stuart
+// Copyright © 2001-2021, Keelan Stuart
 
 
 #pragma once
@@ -16,7 +16,7 @@ namespace c3
 	{
 		#define CRC32_INITVALUE		0xFFFFFFFF
 
-		uint32_t C3_API Calculate(const uint8_t *data, uint32_t len, uint32_t initvalue = CRC32_INITVALUE);
+		uint32_t C3_API Calculate(const uint8_t *data, size_t len, uint32_t initvalue = CRC32_INITVALUE);
 		uint32_t C3_API CalculateString(const TCHAR *s, uint32_t initvalue = CRC32_INITVALUE);
 	};
 
@@ -24,7 +24,7 @@ namespace c3
 	{
 		#define CRC16_INITVALUE		0x0000
 
-		uint16_t C3_API Calculate(const uint8_t *data, uint32_t len, uint16_t initvalue = CRC16_INITVALUE);
+		uint16_t C3_API Calculate(const uint8_t *data, size_t len, uint16_t initvalue = CRC16_INITVALUE);
 		uint16_t C3_API CalculateString(const TCHAR *s, uint16_t initvalue = CRC16_INITVALUE);
 	};
